requirements.txt
streamlit
plotly
snowflake-connector-python[pandas]
scikit-learn
pandas
numpy
python-dotenv
matplotlib